restructure for proper hover view in IDE
