+++ include/numc/internal/core/ndarray.h
