+++ include/numc/internal/utils.h
